+++ src/vsmtp/vsmtp-delivery/src/transport/mbox.rs
@@ -58,6 +58,7 @@ impl Transport for MBox {
                 // NOTE: only linux system is supported here, is the
                 //       path to all mboxes always /var/mail ?
                 write_content_to_mbox(
+                    rcpt,
                     &std::path::PathBuf::from_iter(["/", "var", "mail", rcpt.address.local_part()]),
                     &user,
                     config.server.system.group_local.as_ref(),
@@ -79,8 +80,10 @@ impl Transport for MBox {
                 Some(Err(e)) => {
                     log::error!(
                         target: log_channels::MBOX,
-                        "failed to write email '{}' in mbox of '{rcpt}': {e}",
-                        metadata.message_id
+                        "failed to write email '{}' in {}'s mbox: {}",
+                        metadata.message_id,
+                        rcpt.address.local_part(),
+                        e
                     );
 
                     rcpt.email_status.held_back(e);
@@ -88,8 +91,10 @@ impl Transport for MBox {
                 None => {
                     log::error!(
                         target: log_channels::MBOX,
-                        "failed to write email '{}' in mbox of '{rcpt}': '{rcpt}' is not a user",
-                        metadata.message_id
+                        "failed to write email '{}' in {}'s mbox: '{}' is not a user",
+                        metadata.message_id,
+                        rcpt.address.local_part(),
+                        rcpt.address.local_part(),
                     );
 
                     rcpt.email_status.held_back(TransferErrors::NoSuchMailbox {
@@ -118,6 +123,7 @@ fn build_mbox_message(
 }
 
 fn write_content_to_mbox(
+    rcpt: &Rcpt,
     mbox: &std::path::Path,
     user: &users::User,
     group_local: Option<&users::Group>,
@@ -127,14 +133,13 @@ fn write_content_to_mbox(
     let mut file = std::fs::OpenOptions::new()
         .create(true)
         .append(true)
-        .open(&mbox)
-        .with_context(|| format!("could not open {:?} mbox", mbox))?;
+        .open(&mbox)?;
 
     chown(mbox, Some(user.uid()), group_local.map(users::Group::gid))
         .with_context(|| format!("could not set owner for '{:?}' mbox", mbox))?;
 
-    std::io::Write::write_all(&mut file, content.as_bytes())
-        .with_context(|| format!("could not write email to '{:?}' mbox", mbox))?;
+    std::io::Write::write_all(&mut file, format!("Delivered-To: {rcpt}\n").as_bytes())?;
+    std::io::Write::write_all(&mut file, content.as_bytes())?;
 
     log::debug!(
         target: log_channels::MBOX,
@@ -205,9 +210,15 @@ This is a raw email.
         let metadata = MessageMetadata::default();
 
         std::fs::create_dir_all("./tests/generated/").expect("could not create temporary folders");
-
-        write_content_to_mbox(&mbox, &user, None, &metadata, content)
-            .expect("could not write to mbox");
+        write_content_to_mbox(
+            &Rcpt::new(addr!("john.doe@example.com")),
+            &mbox,
+            &user,
+            None,
+            &metadata,
+            content,
+        )
+        .expect("could not write to mbox");
 
         assert_eq!(
             content.to_string(),